requirements.txt
# Core dependencies
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
requests>=2.31.0

# Web app
flask>=3.0.0
gunicorn>=21.0.0

# Telegram bot
python-telegram-bot>=20.0
